kafka/consumer/requirements.txt
numpy
pandas
requests
scipy
seaborn
python-dotenv
kafka-python-ng
six
psycopg2-binary